requirements.txt
opencv-python~=4.5.2.52
future~=0.18.2
pillow~=8.0.1